refactor(cart): Move getCurrentUserId to controller layer
