+++ src/main/java/mate/academy/bookstore/controller/ShoppingCartController.java
@@ -7,9 +7,12 @@ import lombok.RequiredArgsConstructor;
 import mate.academy.bookstore.dto.cart.CreateCartItemDto;
 import mate.academy.bookstore.dto.cart.ShoppingCartDto;
 import mate.academy.bookstore.dto.cart.UpdateCartItemDto;
+import mate.academy.bookstore.model.User;
 import mate.academy.bookstore.service.ShoppingCartService;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -28,10 +31,17 @@ import org.springframework.web.bind.annotation.RestController;
 public class ShoppingCartController {
     private final ShoppingCartService shoppingCartService;
 
+    private Long getCurrentUserId() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) authentication.getPrincipal();
+        return user.getId();
+    }
+
     @Operation(summary = "Get user's shopping cart")
     @GetMapping
     public ShoppingCartDto getShoppingCart() {
-        return shoppingCartService.getShoppingCart();
+        Long userId = getCurrentUserId();
+        return shoppingCartService.getShoppingCart(userId);
     }
 
     @Operation(summary = "Add a book to the shopping cart")
@@ -39,20 +49,23 @@ public class ShoppingCartController {
     @ResponseStatus(HttpStatus.CREATED)
     public ShoppingCartDto addBookToShoppingCart(
             @RequestBody @Valid CreateCartItemDto requestDto) {
-        return shoppingCartService.addBookToShoppingCart(requestDto);
+        Long userId = getCurrentUserId();
+        return shoppingCartService.addBookToShoppingCart(userId, requestDto);
     }
 
     @Operation(summary = "Update quantity of a book in the shopping cart")
     @PutMapping("/items/{cartItemId}")
     public ShoppingCartDto updateQuantity(@PathVariable Long cartItemId,
                                           @RequestBody @Valid UpdateCartItemDto requestDto) {
-        return shoppingCartService.updateQuantity(cartItemId, requestDto);
+        Long userId = getCurrentUserId();
+        return shoppingCartService.updateQuantity(userId, cartItemId, requestDto);
     }
 
     @Operation(summary = "Remove a book from the shopping cart")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping("/items/{cartItemId}")
     public void removeBookFromShoppingCart(@PathVariable Long cartItemId) {
-        shoppingCartService.removeBookFromShoppingCart(cartItemId);
+        Long userId = getCurrentUserId();
+        shoppingCartService.removeBookFromShoppingCart(userId, cartItemId);
     }
 }
